requirements.txt
# Core dependencies
numpy>=1.21.0
scikit-learn>=1.0.0
pyyaml>=6.0
tqdm>=4.62.0
pydantic>=1.10.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.910
flake8>=4.0.0
requests>=2.31.0
